Rename collector domain to metrics
* Objective

This commit objective is about changing the naming for the collector
domain to metrics.

* Why

It's just a matter of personal preference.

* How

This commit renames the collector domain to metrics
